Fixed bug with music notation - accidentals were not calculated properly when there was only one bar in the staff
git-svn-id: http://svn.powertabs.org/svn/ptedev/trunk@578 5b1460fe-822e-4aa1-ac1b-8a6412c92c8c

+++ source/app/scorearea.h
@@ -88,7 +88,7 @@ protected:
     void drawStdNotation(boost::shared_ptr<const System> system, boost::shared_ptr<const Staff> staff,
                          const StaffData& currentStaffInfo);
 
-    void adjustAccidentals(QMultiMap<double, StdNotationPainter*>& accidentalsMap);
+    void adjustAccidentals(QMultiMap<int, StdNotationPainter*>& accidentalsMap);
     void drawSystemSymbols(Score* score, boost::shared_ptr<const System> system, const StaffData& currentStaffInfo);
     void drawDividerLine(const StaffData& currentStaffInfo, quint32 y);
